Code from R. Kent Dybvig's dissertation for PLT
Implementations of the compilers and VMs from R. Kent Dybvig's "Three Implementation Models for Scheme." Each file is named after the section of the paper it implements. I made a few modifications to make the code work with PLT Scheme. Each file is a module that exports an evaluate function. Some of the VMs need mutable lists, so I explicitly use mcons, etc. where needed. Also, the recur macro is the same as let, so I used let instead.